claude-version/backend/pyproject.toml
[project]
name = "msai"
version = "0.1.0"
description = "MSAI v2 - Personal Hedge Fund Platform"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi>=0.133.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "duckdb>=1.4.0",
    "pyarrow>=22.0.0",
    "pandas>=2.2.0",
    "PyJWT[crypto]>=2.9.0",
    "cryptography>=43.0.0",
    "httpx>=0.28.0",
    "arq>=0.26.0",
    "redis>=5.2.0",
    "nautilus_trader>=1.222.0",
    "quantstats>=0.0.81",
    "ib_async>=1.0.0",
    "databento>=0.43.0",
    "structlog>=24.4.0",
    "typer>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "testcontainers[postgres,redis]>=4.9.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]
azure = [
    "azure-identity>=1.19.0",
    "azure-keyvault-secrets>=4.9.0",
]

[project.scripts]
msai = "msai.cli:app"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
